# Improve rendering speed and fix text alignment issue