--- a/change/@ni-nimble-components-330f85ba-9862-4e5a-8422-20d7a7143a1e.json
+++ b/change/@ni-nimble-components-330f85ba-9862-4e5a-8422-20d7a7143a1e.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Fixed issues with rendering",
+  "packageName": "@ni/nimble-components",
+  "email": "62277788+zszilagy@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}