--- a/packages/nimble-components/src/wafer-map/tests/wafer-map.stories.ts
+++ b/packages/nimble-components/src/wafer-map/tests/wafer-map.stories.ts
@@ -44,7 +44,7 @@ const getDiesSet = (
             returnedValue = sets[1]!;
             break;
         case 'largeGoodSet':
-            returnedValue = generateWaferData(100000, goodValueGenerator(seed))!;
+            returnedValue = generateWaferData(100, goodValueGenerator(seed))!;
             break;
         case 'largeBadSet':
             returnedValue = generateWaferData(100, badValueGenerator(seed))!;